python/samples/getting_started/agents/video_translation_agent/requirements.txt
azure-identity>=1.17.1
azure-storage-blob>=12.22.0
agent-framework[azure] @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/main
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
urllib3>=2.0.0
orjson>=3.9.0
pydantic>=2.0.0
termcolor>=2.3.0
azure-cognitiveservices-speech>=1.40.0
typing-extensions>=4.8.0
nest-asyncio>=1.5.8
azure-core>=1.29.0